Split main class into more files
