+++ Pod/Classes/BetterSegmentedControl+Options.swift
