+++ Pod/Classes/Helpers/UIView+Helpers.swift
@@ -1,0 +1,13 @@
+//
+//  UIView+Helpers.swift
+//  BetterSegmentedControl
+//
+//  Created by George Marmaridis on 19.10.20.
+//
+
+import UIKit
+
+extension UIView {
+    var width: CGFloat { bounds.width }
+    var height: CGFloat { bounds.height }
+}
